Email_Agent/Reply-Handler/requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
pytest
